site refactor implomented
