+++ Client/src/Movie/MovieApp/MovieApp.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import './MovieAppS.scss';
-import api from './key';
 import MovieList from '../MovieList/MovieList';
-
+import api from '../../key';
 
 interface IMovie {
   crew?: string;
@@ -35,30 +34,7 @@ const useMovies = () => {
   return movies;
 }
 
-const useTrailer = (movieId: string | undefined, loadTrailer: boolean) => {
-  const [trailerUrl, setTrailerUrl] = useState<any>({});
-  const [trailerLoaded, setTrailerLoaded] = useState<boolean>(false);
-
-  useEffect( ()=> {
-    async function getRequestTrailer() {
-      const response = await fetch(`https://imdb-api.com/en/API/YouTubeTrailer/${api}/${movieId}`); // random movie
-      const json = await response.json();
-      setTrailerUrl(json);
-      setTrailerLoaded(true);
-    }
-    if (loadTrailer && !trailerLoaded) {
-      getRequestTrailer();
-    }
-
-  },[movieId, loadTrailer, trailerLoaded]);
-
-  useEffect(() => {
-    setTrailerLoaded(false);
-  }, [movieId])
-
-  return trailerUrl;
-
-}
+//usetrailer was here
 
 
 const SandBox: React.FC = () => {
@@ -66,7 +42,7 @@ const SandBox: React.FC = () => {
   const [selectedMovie, setSelectedMovie] = useState<number>(-1);
   const [showTrailer, setShowTrailer] = useState<boolean>(false);
   const [selectedMovieTrailer, setSelectedMovieTrailer] = useState<string | undefined>('')
-  const trailerUrl = useTrailer(selectedMovieTrailer, showTrailer);
+  // const trailerUrl = useTrailer(selectedMovieTrailer, showTrailer);
 
 
 
@@ -95,33 +71,6 @@ const SandBox: React.FC = () => {
           })
         }
       </div>
-      {/* {
-         selectedMovie >= 0 && (
-
-          <div className='moreInfo' style={{display: "flex", flexDirection: "column" }}>
-
-            <h3 className='title'> {title} </h3>
-            <img src={image} width='120rem' height='auto'/>
-            <div className="rating">imDB rating: {imDbRating} </div>
-            <div className='rank'>imDB Rank: {rank} </div>
-            <div className='year'>Year Released: {year} </div>
-            <div className='crew'>Cast/Crew: {crew} </div>
-
-            <button onClick={ ()=>{
-              setSelectedMovieTrailer(id);
-              setShowTrailer(!showTrailer)
-
-            }}> Click for Trailer </button>
-
-            {
-              showTrailer && (
-                <iframe src={`https://www.youtube.com/embed/${trailerUrl.videoId}`}> </iframe>
-              )
-            }
-
-          </div>
-        )
-      } */}
     </div>
   )
