+++ Client/src/Movie/MovieList/MovieListS.scss
@@ -1,0 +1,15 @@
+.movie-main {
+	padding: 10px;
+	display: flex;
+}
+.movie-title-picture-container {
+	flex-direction: row;
+}
+
+.movie-basic-info {
+	flex-direction: column;
+}
+
+.movie-basic-info {
+	padding-left: 15px;
+}
